+++ node/src/components/storage.rs
@@ -70,7 +70,7 @@ use tracing::{debug, error, info, trace, warn};
 use casper_hashing::Digest;
 use casper_types::{
     bytesrepr::{FromBytes, ToBytes},
-    EraId, ExecutionResult, ProtocolVersion, PublicKey, TimeDiff, Timestamp, Transfer, Transform,
+    EraId, ExecutionResult, ProtocolVersion, PublicKey, Timestamp, Transfer, Transform,
 };
 
 use crate::{
@@ -94,7 +94,7 @@ use crate::{
         BlockHashAndHeight, BlockHeader, BlockHeaderWithMetadata, BlockSignatures,
         BlockWithMetadata, Deploy, DeployHash, DeployHeader, DeployId, DeployMetadata,
         DeployMetadataExt, DeployWithFinalizedApprovals, FinalitySignature, FinalizedApprovals,
-        FinalizedBlock, LegacyDeploy, NodeId, SyncLeap, SyncLeapIdentifier, ValueOrChunk,
+        FinalizedBlock, LegacyDeploy, MaxTtl, NodeId, SyncLeap, SyncLeapIdentifier, ValueOrChunk,
     },
     utils::{display_error, WithDir},
     NodeRng,
@@ -214,7 +214,7 @@ pub struct Storage {
     #[data_size(skip)]
     metrics: Option<Metrics>,
     /// The maximum TTL of a deploy.
-    max_ttl: TimeDiff,
+    max_ttl: MaxTtl,
 }
 
 /// A storage component event.
@@ -359,7 +359,7 @@ impl Storage {
         protocol_version: ProtocolVersion,
         activation_era: EraId,
         network_name: &str,
-        max_ttl: TimeDiff,
+        max_ttl: MaxTtl,
         recent_era_count: u64,
         registry: Option<&Registry>,
         force_resync: bool,
@@ -1355,7 +1355,11 @@ impl Storage {
         let timestamp = match self.switch_block_era_id_index.keys().last() {
             Some(era_id) => self
                 .get_switch_block_header_by_era_id(&mut txn, *era_id)?
-                .map(|switch_block| switch_block.timestamp().saturating_sub(self.max_ttl))
+                .map(|switch_block| {
+                    switch_block
+                        .timestamp()
+                        .saturating_sub(self.max_ttl.value())
+                })
                 .unwrap_or_else(Timestamp::now),
             None => Timestamp::now(),
         };
@@ -1592,6 +1596,38 @@ impl Storage {
             )?;
         }
         Ok(true)
+    }
+
+    /// Retrieves single switch block by era ID by looking it up in the index and returning it.
+    fn get_switch_block_by_era_id<Tx: Transaction>(
+        &self,
+        txn: &mut Tx,
+        era_id: EraId,
+    ) -> Result<Option<Block>, FatalStorageError> {
+        self.switch_block_era_id_index
+            .get(&era_id)
+            .and_then(|block_hash| self.get_single_block(txn, block_hash).transpose())
+            .transpose()
+    }
+
+    /// Get the switch block for a specified era number in a read-only LMDB database transaction.
+    ///
+    /// # Panics
+    ///
+    /// Panics on any IO or db corruption error.
+    pub(crate) fn read_switch_block_by_era_id(
+        &self,
+        era_id: EraId,
+    ) -> Result<Option<Block>, FatalStorageError> {
+        let mut txn = self
+            .env
+            .begin_ro_txn()
+            .expect("Could not start read only transaction for lmdb");
+        let switch_block = self
+            .get_switch_block_by_era_id(&mut txn, era_id)
+            .expect("LMDB panicked trying to get switch block");
+        txn.commit().expect("Could not commit transaction");
+        Ok(switch_block)
     }
 
     /// Returns `count` highest switch block headers, sorted from lowest (oldest) to highest.
@@ -2419,7 +2455,7 @@ impl Storage {
         }
     }
 
-    fn get_available_block_range(&self) -> AvailableBlockRange {
+    pub(crate) fn get_available_block_range(&self) -> AvailableBlockRange {
         match self.completed_blocks.highest_sequence() {
             Some(&seq) => seq.into(),
             None => AvailableBlockRange::RANGE_0_0,
@@ -2856,38 +2892,6 @@ impl Storage {
                 DeployHash::new(Digest::try_from(raw_key).expect("malformed deploy hash in DB"))
             })
             .collect()
-    }
-
-    /// Retrieves single switch block by era ID by looking it up in the index and returning it.
-    fn get_switch_block_by_era_id<Tx: Transaction>(
-        &self,
-        txn: &mut Tx,
-        era_id: EraId,
-    ) -> Result<Option<Block>, FatalStorageError> {
-        self.switch_block_era_id_index
-            .get(&era_id)
-            .and_then(|block_hash| self.get_single_block(txn, block_hash).transpose())
-            .transpose()
-    }
-
-    /// Get the switch block for a specified era number in a read-only LMDB database transaction.
-    ///
-    /// # Panics
-    ///
-    /// Panics on any IO or db corruption error.
-    pub(crate) fn transactional_get_switch_block_by_era_id(
-        &self,
-        switch_block_era_num: u64,
-    ) -> Result<Option<Block>, FatalStorageError> {
-        let mut txn = self
-            .env
-            .begin_ro_txn()
-            .expect("Could not start read only transaction for lmdb");
-        let switch_block = self
-            .get_switch_block_by_era_id(&mut txn, EraId::from(switch_block_era_num))
-            .expect("LMDB panicked trying to get switch block");
-        txn.commit().expect("Could not commit transaction");
-        Ok(switch_block)
     }
 
     /// Directly returns a deploy from internal store.