+++ utils/nctl/sh/scenarios/common/itst.sh
@@ -225,6 +225,14 @@ function do_await_era_change() {
     nctl-await-n-eras offset="$ERA_COUNT" sleep_interval='5.0'
 }
 
+function do_await_era_change_with_timeout() {
+    # allow chain height to grow
+    local ERA_COUNT=${1:-"1"}
+    local TIME_OUT=${2:-''}
+    log_step "awaiting $ERA_COUNT eras…"
+    nctl-await-n-eras offset="$ERA_COUNT" sleep_interval='5.0' timeout="$TIME_OUT"
+}
+
 function check_current_era {
     local NODE_ID=${1:-$(get_node_for_dispatch)}
     local ERA="null"
@@ -579,4 +587,29 @@ function delegate_to() {
         amount="$AMOUNT" \
         delegator="$ACCOUNT_ID" \
         validator="$NODE_ID"
+}
+
+function assert_chain_stalled() {
+    # Fucntion checks that the two remaining node's LFB checked
+    # n-seconds apart doesnt progress
+    log_step "ensuring chain stalled"
+    local SLEEP_TIME=${1}
+    # Sleep 5 seconds to allow for final message propagation.
+    sleep 5
+    local LFB_1_PRE=$(do_read_lfb_hash 1)
+    local LFB_2_PRE=$(do_read_lfb_hash 2)
+    log "Sleeping ${SLEEP_TIME}s..."
+    sleep $SLEEP_TIME
+    local LFB_1_POST=$(do_read_lfb_hash 1)
+    local LFB_2_POST=$(do_read_lfb_hash 2)
+
+    if [ "$LFB_1_PRE" != "$LFB_1_POST" ] && [ "$LFB_2_PRE" != "$LFB_2_POST" ]; then
+       log "Error: Chain progressed."
+       exit 1
+    else
+        STALLED_LFB=$LFB_1_POST
+        log "node-1 LFB: $LFB_1_PRE = $LFB_1_POST"
+        log "node-2 LFB: $LFB_2_PRE = $LFB_2_POST"
+        log "Stall successfully detected, continuing..."
+    fi
 }